separated components from features/map/index
