+++ src/features/map/map-padding.js
@@ -1,40 +1,13 @@
 import React from 'react';
 
-import { getTileSprite, FogTile } from './index.js';
+import { getTileSprite, FogTile } from './map-tile';
 import { SPRITE_SIZE }            from '../../config/constants';
 
 import './styles.scss';
 
-function BoundaryTile({ tileType, variation, explored, sightBox, location }) {
+const MapPadding = ({ tileType, tiles, sightBox }) => {
 
-  let inSight = false;
-
-  if(sightBox) {
-    // check the sight box tiles
-    sightBox.forEach(sightBoxTile => {
-      // if the current tile is in range
-      if(JSON.stringify(sightBoxTile) === JSON.stringify(location)) {
-        // remove the overlay
-        return inSight = true;
-      }
-    });
-  }
-
-  return(
-    <div style={{
-      backgroundImage: `url(/tiles/${getTileSprite(tileType, variation)}.png)`,
-      display: 'inline-flex',
-      height: SPRITE_SIZE,
-      width: SPRITE_SIZE
-    }}>
-      <FogTile explored={explored} inSight={inSight} />
-    </div>
-  );
-}
-
-export default function MapPadding({ tileType, tiles, sightBox }) {
-
-  const PaddingTiles = {};
+  let PaddingTiles = {};
 
   Object.keys(tiles).forEach(direction => {
     PaddingTiles[direction] = tiles[direction].map((row, index) => {
@@ -79,3 +52,32 @@ export default function MapPadding({ tileType, tiles, sightBox }) {
     </>
   );
 }
+
+function BoundaryTile({ tileType, variation, explored, sightBox, location }) {
+
+  let inSight = false;
+
+  if(sightBox) {
+    // check the sight box tiles
+    sightBox.forEach(sightBoxTile => {
+      // if the current tile is in range
+      if(JSON.stringify(sightBoxTile) === JSON.stringify(location)) {
+        // remove the overlay
+        return inSight = true;
+      }
+    });
+  }
+
+  return(
+    <div style={{
+      backgroundImage: `url(/tiles/${getTileSprite(tileType, variation)}.png)`,
+      display: 'inline-flex',
+      height: SPRITE_SIZE,
+      width: SPRITE_SIZE
+    }}>
+      <FogTile explored={explored} inSight={inSight} />
+    </div>
+  );
+}
+
+export default MapPadding;
