+++ src/features/player/actions/explore-tiles.js
@@ -2,9 +2,9 @@ import getSurroundingTiles from '../../../utils/get-surrounding-tiles';
 
 export default function exploreTiles(newPos) {
   return dispatch => {
+
     const { tiles, paddingTiles } = getSurroundingTiles(newPos);
 
-    // explore the tiles
     dispatch({
       type: 'EXPLORE_TILES',
       payload: {
